requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy>=2.0.0
passlib>=1.7.4
bcrypt==3.2.2
